Seperate upload code from build-project.py
